Locust/requirements.txt
confluent-kafka==2.6.0
locust-plugins==4.5.3